Fix types of opts

+++ src/emGUI/options/opts.h
@@ -32,50 +32,50 @@ typedef EMGUI_XFONT_TYPE			xFont; //TODO: Add GFXFont segmented struct to allow 
 #endif
 
 // Dimensions
-extern int EMGUI_LCD_WIDTH;
-extern int EMGUI_LCD_HEIGHT;
-extern int EMGUI_STATUS_BAR_HEIGHT; 
-extern int EMGUI_MODAL_DLG_BTN_SPACING;
+extern uint16_t EMGUI_LCD_WIDTH;
+extern uint16_t EMGUI_LCD_HEIGHT;
+extern uint16_t EMGUI_STATUS_BAR_HEIGHT; 
+extern uint16_t EMGUI_MODAL_DLG_BTN_SPACING;
 
 // String limits
-extern int EMGUI_WINDOW_HEADER_LENGTH;
-extern int EMGUI_LABEL_MAX_LENGTH;
-extern int EMGUI_DEBUG;
+extern uint16_t EMGUI_WINDOW_HEADER_LENGTH;
+extern uint16_t EMGUI_LABEL_MAX_LENGTH;
+extern uint16_t EMGUI_DEBUG;
 
 // GUI COLORS
 
 // RGB565 definition 
-extern int EMGUI_COLOR_SEA_5; 		// Blue colors map from drakest to lightest
-extern int EMGUI_COLOR_SEA_4;
-extern int EMGUI_COLOR_SEA_3;
-extern int EMGUI_COLOR_SEA_2;
-extern int EMGUI_COLOR_SEA_1;
+extern uint16_t EMGUI_COLOR_SEA_5; 		// Blue colors map from drakest to lightest
+extern uint16_t EMGUI_COLOR_SEA_4;
+extern uint16_t EMGUI_COLOR_SEA_3;
+extern uint16_t EMGUI_COLOR_SEA_2;
+extern uint16_t EMGUI_COLOR_SEA_1;
 
-extern int EMGUI_COLOR_BROWN_5;	// Brown colors map from drakest to lightest
-extern int EMGUI_COLOR_BROWN_4;
-extern int EMGUI_COLOR_BROWN_3;
-extern int EMGUI_COLOR_BROWN_2;
-extern int EMGUI_COLOR_BROWN_1;
+extern uint16_t EMGUI_COLOR_BROWN_5;	// Brown colors map from drakest to lightest
+extern uint16_t EMGUI_COLOR_BROWN_4;
+extern uint16_t EMGUI_COLOR_BROWN_3;
+extern uint16_t EMGUI_COLOR_BROWN_2;
+extern uint16_t EMGUI_COLOR_BROWN_1;
 
-extern int EMGUI_COLOR_WHITE;
-extern int EMGUI_COLOR_GRAY;
-extern int EMGUI_COLOR_BLACK;
+extern uint16_t EMGUI_COLOR_WHITE;
+extern uint16_t EMGUI_COLOR_GRAY;
+extern uint16_t EMGUI_COLOR_BLACK;
 
-extern int EMGUI_COLOR_RED;
-extern int EMGUI_COLOR_GREEN;
-extern int EMGUI_COLOR_BLUE;
+extern uint16_t EMGUI_COLOR_RED;
+extern uint16_t EMGUI_COLOR_GREEN;
+extern uint16_t EMGUI_COLOR_BLUE;
 
 
 // STATUSBAR STORY
-extern int EMGUI_COLOR_MENU_HEADER_TEXT; 
-extern int EMGUI_COLOR_STATUS_BAR_BG;
-extern int EMGUI_COLOR_MESSAGE_TEXT;
-extern int EMGUI_MENU_BUTTON_LABEL_TEXT_COLOR;
-extern int EMGUI_MENU_BUTTON_LABEL_BG_COLOR;
+extern uint16_t EMGUI_COLOR_MENU_HEADER_TEXT; 
+extern uint16_t EMGUI_COLOR_STATUS_BAR_BG;
+extern uint16_t EMGUI_COLOR_MESSAGE_TEXT;
+extern uint16_t EMGUI_MENU_BUTTON_LABEL_TEXT_COLOR;
+extern uint16_t EMGUI_MENU_BUTTON_LABEL_BG_COLOR;
 
 // PLOT Colors
-extern int EMGUI_COLOR_PLOT_BACKGROUND;
-extern int EMGUI_COLOR_PLOT_GRIDS;
-extern int EMGUI_COLOR_PLOT_SCALE_MARKER;
+extern uint16_t EMGUI_COLOR_PLOT_BACKGROUND;
+extern uint16_t EMGUI_COLOR_PLOT_GRIDS;
+extern uint16_t EMGUI_COLOR_PLOT_SCALE_MARKER;
 
 #endif // !OPTS_H
